Contact page, css
Filled in contact details on contact page. Wrote a piece of additional
text for it some time ago, to add that in next time.

Made minor styling changes in css. *Dreamweaver's live preview on my
laptop is significantly smaller than what actually shows in the
published ver, need to keep in mind.