CSS
Small but important changes made in css, applying to about us page.

Current issue is with the text spanning the whole width of the browser.
Several possible options to fix this, including enclosing text with
tables or articles. The viability of these options depends on how well
they translate in the css, so is a major issue.
>To check other examples of how text 'frames' are handled.

(Following up on earlier note on Dreamweaver's inaccurate in-app live
preview- opening the pages, eg contact.html from file explorer into
browser resolves this problem.)